Extract base class from StateSender
