+++ custom_components/ta_coe/__init__.py
@@ -25,7 +25,7 @@ from .const import (
 from .issues import check_coe_server_2x_issue
 from .refresh_task import RefreshTask
 from .state_observer import StateObserver
-from .state_sender import StateSender
+from .state_sender_v1 import StateSenderV1
 
 PLATFORMS: list[str] = [Platform.SENSOR, Platform.BINARY_SENSOR]
 
@@ -52,7 +52,7 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 
     await coordinator.async_config_entry_first_refresh()
 
-    sender = StateSender(coe, entry.data.get(CONF_ENTITIES_TO_SEND, {}))
+    sender = StateSenderV1(coe, entry.data.get(CONF_ENTITIES_TO_SEND, {}))
     observer = StateObserver(
         hass, coe, sender, entry.data.get(CONF_ENTITIES_TO_SEND, {})
     )
