+++ tests/test_state_sender.py
@@ -6,9 +6,9 @@ import pytest
 from ta_cmi import CoE, CoEChannel
 from ta_cmi.const import ChannelMode
 
-from custom_components.ta_coe import StateSender
+from custom_components.ta_coe import StateSenderV1
 from custom_components.ta_coe.const import ANALOG_DOMAINS, DIGITAL_DOMAINS
-from custom_components.ta_coe.state_sender import AnalogValue
+from custom_components.ta_coe.state_sender_v1 import AnalogValue
 from tests import COE_SEND_ANALOG_VALUES_PACKAGE, COE_SEND_DIGITAL_VALUES_PACKAGE
 
 coe = CoE("")
@@ -28,7 +28,7 @@ def test_sender_init_create_digital_pages():
     """When sender is initialized then empty digital states for all entities are created."""
     count = 8
 
-    sender = StateSender(coe, create_dummy_ids(DIGITAL_DOMAINS[0], count))
+    sender = StateSenderV1(coe, create_dummy_ids(DIGITAL_DOMAINS[0], count))
 
     assert len(sender._digital_states) == count
 
@@ -40,7 +40,7 @@ def test_sender_init_create_analog_pages():
     """When sender is initialized then empty analog states for all entities are created."""
     count = 8
 
-    sender = StateSender(coe, create_dummy_ids(ANALOG_DOMAINS[0], count))
+    sender = StateSenderV1(coe, create_dummy_ids(ANALOG_DOMAINS[0], count))
 
     assert len(sender._analog_states) == count
 
@@ -57,7 +57,7 @@ def test_sender_init_create_map_entity_id_to_index():
         "3": "binary_sensor.4",
     }
 
-    sender = StateSender(coe, entity_ids)
+    sender = StateSenderV1(coe, entity_ids)
 
     assert sender._index_from_id == {
         entity_ids["0"]: "0",
@@ -70,7 +70,7 @@ def test_sender_init_create_map_entity_id_to_index():
 def test_sender_digital_manuel_change_digital_page():
     """Test when state manuel added then the state on position is updated."""
     entities_ids = create_dummy_ids(DIGITAL_DOMAINS[0], 8)
-    sender = StateSender(coe, entities_ids)
+    sender = StateSenderV1(coe, entities_ids)
 
     index_to_change = 1
 
@@ -83,7 +83,7 @@ def test_sender_digital_manuel_change_digital_page():
 def test_sender_analog_manuel_change_analog_page():
     """Test when state manuel added then the state on position is updated."""
     entities_ids = create_dummy_ids(ANALOG_DOMAINS[0], 8)
-    sender = StateSender(coe, entities_ids)
+    sender = StateSenderV1(coe, entities_ids)
 
     index_to_change = 1
 
@@ -105,7 +105,7 @@ def test_sender_analog_manuel_change_analog_page():
 async def test_sender_digital_change_digital_page(page: int):
     """Test when state added then the state on position is updated and changes are send to the server."""
     entities_ids = create_dummy_ids(DIGITAL_DOMAINS[0], 18)
-    sender = StateSender(coe, entities_ids)
+    sender = StateSenderV1(coe, entities_ids)
 
     index_to_change = 16 * (page - 1)
 
@@ -128,7 +128,7 @@ async def test_sender_digital_change_digital_page(page: int):
 async def test_sender_analog_change_analog_page(page: int):
     """Test when state added then the state on position is updated and changes are send to the server."""
     entities_ids = create_dummy_ids(ANALOG_DOMAINS[0], 30)
-    sender = StateSender(coe, entities_ids)
+    sender = StateSenderV1(coe, entities_ids)
 
     index_to_change = 4 * (page - 1)
 
@@ -159,7 +159,7 @@ async def test_sender_analog_change_analog_page(page: int):
 async def test_sender_update_digital():
     """Test when updated all states are send to the server. Test with digital states."""
     entities_ids = create_dummy_ids(DIGITAL_DOMAINS[0], 30)
-    sender = StateSender(coe, entities_ids)
+    sender = StateSenderV1(coe, entities_ids)
 
     with patch(COE_SEND_DIGITAL_VALUES_PACKAGE) as update_mock, patch(
         COE_SEND_ANALOG_VALUES_PACKAGE
@@ -188,7 +188,7 @@ async def test_sender_update_digital():
 async def test_sender_update_analog():
     """Test when updated all states are send to the server. Test with analog states."""
     entities_ids = create_dummy_ids(ANALOG_DOMAINS[0], 30)
-    sender = StateSender(coe, entities_ids)
+    sender = StateSenderV1(coe, entities_ids)
 
     with patch(COE_SEND_ANALOG_VALUES_PACKAGE) as update_mock, patch(
         COE_SEND_DIGITAL_VALUES_PACKAGE
